Updates to grind75 and add similar for neetcode
Added subcommand to grind75 to print all problems that are not checked.
The neetcode script is similar to grind75 in this regard.
Both grind75 and neetcode a subcommand to print the first unchecked.
With no arguments, they both print all problems.

Also added usage function to both scripts so this is slightly more
discoverable.